requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==4.2.13
django-cors-headers==4.2.0
django-jazzmin==2.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
frozenlist==1.4.1
idna==3.7
inflection==0.5.1
multidict==6.0.5
packaging==24.1
pillow==10.3.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
sqlparse==0.5.0
telepot==12.7

